+++ digiwf-okewo-integration-service/src/main/java/io/muenchendigital/digiwf/okewo/integration/api/dto/request/OrdnungsmerkmalDto.java
@@ -1,4 +1,4 @@
-package io.muenchendigital.digiwf.okewo.integration.api.dto;
+package io.muenchendigital.digiwf.okewo.integration.api.dto.request;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
