+++ digiwf-okewo-integration-service/src/main/java/io/muenchendigital/digiwf/okewo/integration/api/dto/request/SearchPersonErweitertRequestDto.java
@@ -1,4 +1,4 @@
-package io.muenchendigital.digiwf.okewo.integration.api.dto;
+package io.muenchendigital.digiwf.okewo.integration.api.dto.request;
 
 import io.muenchendigital.digiwf.okewo.integration.gen.model.SuchePersonerweitertAnfrage;
 import lombok.Data;
